+++ src/main/java/me/marlon/ecs/MeshSystem.java
@@ -9,7 +9,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 public class MeshSystem implements IComponentListener, IUpdateListener {
-    private static final short BITS = EntityManager.MESH_BIT | EntityManager.TRANSFORM_BIT;
+    private static final int BITS = EntityManager.MESH_BIT | EntityManager.TRANSFORM_BIT;
 
     private EntityManager entities;
     private Renderer renderer;
